# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LeagueVerse Fantasy Football Draft API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leagues/{leagueId}/teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List teams in a league
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Add a team to a league
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /leagues/{leagueId}/teams/{teamId}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get a team
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update a team
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
...
  /leagues/{leagueId}/teams/{teamId}/needs:
    get:
      operationId: getTeamNeeds
      tags: [teams]
      summary: Get team positional needs analysis
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamNeeds"
  /leagues/{leagueId}/teams/{teamId}/roster:
    get:
      operationId: getTeamRoster
      tags: [teams]
      summary: Get a team's current roster
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pick"
...
components:
  schemas:
...
    Team:
      type: object
      required: [id, leagueId, name, ownerName, draftPosition, createdAt]
      properties:
        id:
          type: integer
        leagueId:
          type: integer
        name:
          type: string
        ownerName:
          type: string
        draftPosition:
          type: integer
        logoUrl:
          type: ["string", "null"]
        walkUpSong:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        createdAt:
          type: string
    TeamInput:
      type: object
      required: [name, ownerName, draftPosition]
      properties:
        name:
          type: string
        ownerName:
          type: string
        draftPosition:
          type: integer
        logoUrl:
          type: string
        walkUpSong:
          type: string
        primaryColor:
          type: string
    TeamUpdate:
      type: object
      properties:
        name:
          type: string
        ownerName:
          type: string
        logoUrl:
          type: string
        walkUpSong:
          type: string
        primaryColor:
          type: string
    Player:
      type: object
      required: [id, name, position, nflTeam, rank, adp, projectedPoints, byeWeek, status]
      properties:
        id:
          type: integer
        name:
          type: string
        position:
          type: string
          enum: [QB, RB, WR, TE, K, DEF]
        nflTeam:
          type: string
        rank:
          type: integer
        adp:
          type: number
        projectedPoints:
          type: number
        byeWeek:
          type: integer
        status:
          type: string
          enum: [active, injured, questionable, out, ir]
        injuryNote:
          type: ["string", "null"]
        isDrafted:
          type: boolean
    Pick:
      type: object
      required: [id, leagueId, teamId, playerId, round, pickInRound, overallPick, createdAt]
      properties:
        id:
          type: integer
        leagueId:
          type: integer
        teamId:
          type: integer
        playerId:
          type: integer
        round:
          type: integer
        pickInRound:
          type: integer
        overallPick:
          type: integer
        grade:
          type: ["string", "null"]
        gradeExplanation:
          type: ["string", "null"]
        valueScore:
          type: ["number", "null"]
        createdAt:
          type: string
        player:
          $ref: "#/components/schemas/Player"
        team:
          $ref: "#/components/schemas/Team"
...
    TeamNeeds:
      type: object
      required: [teamId, needs]
      properties:
        teamId:
          type: integer
        needs:
          type: array
          items:
            $ref: "#/components/schemas/PositionNeed"
    PositionNeed:
      type: object
      required: [position, needLevel, currentCount, targetCount, message]
      properties:
        position:
          type: string
        needLevel:
          type: string
          enum: [none, low, medium, high, critical]
        currentCount:
          type: integer
        targetCount:
          type: integer
        message:
          type: string
...